Task 1.2 finished

+++ week1/class_utils.py
@@ -3,10 +3,10 @@ from dataclasses import dataclass
 
 @dataclass
 class BoundingBox:
-    x1: int
-    y1: int
-    x2: int
-    y2: int
+    x1: float
+    y1: float
+    x2: float
+    y2: float
     track_id: int
     frame: int
     label: str
